requirements.txt
streamlit>=1.0
pandas>=1.3
numpy>=1.21
plotly>=5.0
seaborn>=0.11
matplotlib>=3.4
openai>=1.0.0
# Providers / env support
google-generativeai>=0.8.0
python-dotenv>=1.0
# Excel readers
openpyxl>=3.1  # for .xlsx files
xlrd>=2.0      # for .xls files
# Optional forecasting (heavy install); uncomment if needed
# prophet>=1.1
reportlab>=3.6
# Optional: for PDF rendering via wkhtmltopdf + pdfkit
# pdfkit>=1.0
# Optional: local LLM client
# ollama>=0.1.0
